Create Clickerman_UDLs
Clickermann is a great scripting autoclicker program and I have been using it for many years.

Clickermann site:
http://crapware.aidf.org/
... and forum:
http://crapware.aidf.org/forum/

Unfortunately, the author of Clickermann has not released English version of the application, and you should use it in Russian. 

-----------------------------------------------------------------------

Clickermann has its own code editor, but any large code is usually split into a few (or many) files. Since Clickermann doesn't support tabs in the editor, you have to switch between files frequently, which is very slow. So I use Notepad++ to view and quickly edit parts of the code in a separate file. To improve the readability of the code, I made UDLs for the light theme and for the dark theme.